Create CalculatiorTest Module
This module allows the bod to take in a simple operation, only including addition, subtraction, multiplication and division. The module only allows for one calculation per command and if the format is not followed, (number operation number) the bot will say so and ask you to reenter your command using the correct format. This simple calculator can deal with both decimals and negative numbers, as all numbers are stored as floats